requirements.txt
psycopg2-binary==2.7.5
pyasn1==0.4.4
pyasn1-modules==0.2.2
python-ldap==3.1.0
